chore: initial directory restructure
