+++ ai_giphy/app/activities.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict
 
 # Assuming ai_agent.py is in the same directory
-from ai_agent import get_chain
+from app.ai_agent import get_chain
 from application_sdk.activities import ActivitiesInterface
 from application_sdk.observability.logger_adaptor import get_logger
 from temporalio import activity
