+++ ai_giphy/pyproject.toml
@@ -1,41 +1,24 @@
 [project]
-name = "atlan-sample-apps"
+name = "ai-giphy-app"
 version = "0.1.0"
-description = "Atlan Sample Apps: Sample apps built using Atlan Application SDK"
+description = "AI Giphy sample app built using Atlan Application SDK"
 authors = [{ name = "Atlan App Team", email = "connect@atlan.com" }]
 requires-python = ">=3.11"
 license = "Apache-2.0"
 readme = "README.md"
 dependencies = [
-    "atlan-application-sdk[workflows,sqlalchemy,tests,daft]==0.1.1rc23",
-    "poethepoet"
-]
-
-# If you have development dependencies, you can add them like this:
-[dependency-groups]
-ai_giphy = [
+    "atlan-application-sdk[workflows,tests]==0.1.1rc23",
+    "poethepoet",
     "langchain>=0.3.25",
     "langchain-openai>=0.3.16",
     "langchainhub>=0.1.21",
 ]
+
+[dependency-groups]
 dev = [
-    "pre-commit>=4.2.0,<5.0.0",
-    "requests"
-]
-mysql = [
-    "pymysql>=1.1.1",
-]
-freshness_monitor = [
-    "pyatlan",
-]
-test = [
     "pytest",
-    "pytest-asyncio",
+    "pytest-asyncio", 
     "coverage",
-]
-asset_description_reminder = [
-    "slack-sdk",
-    "pyatlan",
 ]
 
 [tool.poe.tasks]
@@ -61,14 +44,18 @@ response.raise_for_status()
 for file_info in response.json():
     if file_info["type"] == "file" and file_info["name"].endswith(".yaml"):
         raw_url = file_info["download_url"]
-
+        
         file_response = requests.get(raw_url)
         file_response.raise_for_status()
-
+        
         file_path = components_dir / file_info["name"]
         file_path.write_text(file_response.text)
         print(f"Downloaded: {file_info['name']}")
 """
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.sdist]
 include = ["./**/*.py"]
@@ -76,17 +63,10 @@ include = ["./**/*.py"]
 [tool.hatch.build.targets.wheel]
 include = ["./**/*.py"]
 packages = [
-    "hello_world",
-    "giphy",
-    "ai_giphy",
-    "mysql"
+    "app"
 ]
 
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-
+# for local development
 [tool.uv.sources]
 # atlan-application-sdk = { path = "../application-sdk", editable = true }
 # atlan-application-sdk = { git = "https://github.com/atlanhq/application-sdk", rev = "349e45a1ee273315a4474f645332e2ef4c8569a9" }
