+++ asset_descriptor_reminder/app/client.py
