+++ giphy/app/activities.py
