+++ hello_world/main.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from activities import HelloWorldActivities
+from app.activities import HelloWorldActivities
 from application_sdk.application import BaseApplication
 from application_sdk.observability.decorators.observability_decorator import (
     observability,
@@ -8,7 +8,7 @@ from application_sdk.observability.decorators.observability_decorator import (
 from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.observability.metrics_adaptor import get_metrics
 from application_sdk.observability.traces_adaptor import get_traces
-from workflow import HelloWorldWorkflow
+from app.workflow import HelloWorldWorkflow
 
 logger = get_logger(__name__)
 metrics = get_metrics()
