+++ mysql/main.py
@@ -19,10 +19,10 @@ from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.observability.metrics_adaptor import get_metrics
 from application_sdk.observability.traces_adaptor import get_traces
 
-from mysql.app.activities import SQLMetadataExtractionActivities
-from mysql.app.clients import SQLClient
-from mysql.app.transformer import SQLAtlasTransformer
-from mysql.app.workflows import SQLMetadataExtractionWorkflow
+from app.activities import SQLMetadataExtractionActivities
+from app.clients import SQLClient
+from app.transformer import SQLAtlasTransformer
+from app.workflows import SQLMetadataExtractionWorkflow
 
 logger = get_logger(__name__)
 metrics = get_metrics()
